fix: Eliminate all 91 compiler warnings and enforce brutal quality gates
## Summary
- Fixed all 91 compiler warnings across the entire codebase (now 0)
- Updated app-release-manager to BRUTAL MODE with zero tolerance
- Fixed test_projects.c double-free bug and incorrect return type check
- Added comprehensive test stubs for isolated unit testing

## Warning Fixes (by category)
- Sign conversion: Cast pointer subtraction to (size_t)
- Double promotion: Cast floats to (double) in printf
- Format-nonliteral: Pragma for compile-time format strings
- Shorten-64-to-32: Proper size_t casts for snprintf returns
- Terminal flags: Cast to (tcflag_t) for bitwise operations

## Files Changed (60 total)
- 57 source files with warning fixes
- app-release-manager.md: BRUTAL MODE enabled by default
- test_projects.c: Fixed double-free and added stubs
- Makefile: Simplified test dependencies

## Quality Improvements
- All 154 tests pass (fuzz, unit, compaction, compare, projects)
- Zero compiler warnings with -Wall -Wextra -Werror
- Release manager now blocks on ANY issue (no warnings/minor status)

+++ src/compare/diff.c
@@ -39,7 +39,7 @@ static LineArray split_lines(const char* text) {
 
     for (const char* p = text; *p; p++) {
         if (*p == '\n') {
-            size_t len = p - start;
+            size_t len = (size_t)(p - start);
             arr.lines[idx] = malloc(len + 1);
             if (arr.lines[idx]) {
                 memcpy(arr.lines[idx], start, len);
@@ -197,21 +197,21 @@ char* generate_response_diff(const char* response1, const char* response2,
     }
 
     size_t pos = 0;
-    pos += snprintf(output + pos, output_size - pos, "--- %s\n", label1 ? label1 : "Response 1");
-    pos += snprintf(output + pos, output_size - pos, "+++ %s\n", label2 ? label2 : "Response 2");
+    pos += (size_t)snprintf(output + pos, output_size - pos, "--- %s\n", label1 ? label1 : "Response 1");
+    pos += (size_t)snprintf(output + pos, output_size - pos, "+++ %s\n", label2 ? label2 : "Response 2");
 
     for (size_t i = 0; i < diff_count && pos < output_size - 100; i++) {
         switch (diffs[i].type) {
             case DIFF_SAME:
-                pos += snprintf(output + pos, output_size - pos, "  %s\n",
+                pos += (size_t)snprintf(output + pos, output_size - pos, "  %s\n",
                                diffs[i].text ? diffs[i].text : "");
                 break;
             case DIFF_DELETE:
-                pos += snprintf(output + pos, output_size - pos, "- %s\n",
+                pos += (size_t)snprintf(output + pos, output_size - pos, "- %s\n",
                                diffs[i].text ? diffs[i].text : "");
                 break;
             case DIFF_ADD:
-                pos += snprintf(output + pos, output_size - pos, "+ %s\n",
+                pos += (size_t)snprintf(output + pos, output_size - pos, "+ %s\n",
                                diffs[i].text ? diffs[i].text : "");
                 break;
             case DIFF_CHANGE:
